When restoring the Visual selection, don't contract region if end is bolp.
Since the selection is restored on the basis of the expanded values of
point and mark, we need to contract it. However,
`vimpulse-visual-mark' and `vimpulse-visual-point' may get "unnatural"
positions if text is deleted: for example, the end of the selection
would never occur at the very beginning of a line, since the first
character would be included in the selection. In such a case,
contraction would be errorneous and should be avoided.