Don't convert keystrokes to digits unless the keystroke is a character
`vimpulse-keypress-parser', which reads keystrokes in Operator-Pending
mode, converts keystrokes like "C-0" and "C-M-1" to numerical form
(0 and 1). This conversion is of course not applicable on a keystroke
like <S-return>; in fact, it causes an error since the conversion is
arithmetical and <S-return> can't be represented as an integer
(character). This commit therefore makes the conversion contingent
on `characterp'.