Bugfix: "Vd" doesn't include newline when Vimpulse is byte-compiled
When working with linewise selections, it is often useful to exclude
the final newline (e.g., skeletons). Operator commands, however,
should always receive the "full range". `vimpulse-operator-cmd-p'
inspects the `interactive-form' of the command to be executed, but
fails when the form is byte-compiled. The problem is fixed by
allowing `vimpulse-memq-recursive' to search inside vector elements.