Arrow keys don't work as motions
I also mentioned another problem, that "y2<right>" didn't work, that
I've since managed to fix. The bug was in `vimpulse-keypress-parser':
there's a variable `digit' that gets set when you type a prefix
argument (2 in this case). It's supposed to get reset once you type
the next character, but this only happens if the next character is an
ASCII character, and as a result, if you type in a non-ASCII
character, it gets accidentally treated as a prefix argument instead.