Ensure state bindings for minor modes are properly refreshed
When a minor mode with state bindings is enabled, we should
immediately call `viper-normalize-minor-mode-map-alist' to ensure the
relevant keymaps are ranked properly and `viper-set-mode-vars-for' to
ensure they are active. Otherwise, the state bindings will not work
until the user enters (and exits) Insert state. We let
`vimpulse-define-key' advise the mode function, so that the mentioned
functions are called automatically. Thus, M-x mode or (mode 1) will
automatically refresh the bindings; (setq mode t) will not, however.