Fix `vimpulse-change' bug when point is outside range
Some text objects, such as "i(", may select text that lies ahead of
point. When a "ci(" change is repeated and point is at an opening
parenthesis, the text is inserted before the parentheses instead of
between them. The problem is fixed by ensuring that `vimpulse-change'
always moves point to the text range's beginning.