Remove "whole sentences" rule and reformat comments
Short comments may be uncapitalized and period-less:

    ;; this is a short comment

Longer comments may be neither:

    ;; This is a longer comment.

Until now, "phrase comments" (the first type) were disallowed for
consistency. This strictness came at the price of having to "correct"
many contributions. As it is less work to simply go with common
practices, it makes better sense to lift the restriction.