Fix long-standing compilation issues with the test framework
To let `defsuite' be wrapped around `deftest', there were two
strategies for making the two "communicate": either at compile time or
at run time. The former was tried first, with partial results: tests
would work uncompiled and the file would compile without warnings --
but the tests wouldn't work compiled.

In this commit, we make the communication happen at run time instead.
The macro expansion of `defsuite' is a `let' form which the expansion
of `deftest', running inside the form, can pick up on. The tests now
compile properly, with (ahem) considerable speed gains.

We also try to disentangle the confusing relationship between suites
and tests. Since the current suite is always included in the failure
report, suites, not tests, are charged with the task of reporting
failures. A test function consists of two clauses: either the test
will call a suite to run itself (and report any failure), or it will
execute its test code and return the result. If the test succeeds, the
return value is t; otherwise it is an error message string, which the
calling suite collects and reports.

Any suite can be called to run any test. If a test is not associated
with a suite, it may default to `empty-suite'.