Make "*" and "#" preserve search when pressed repeatedly
If we have the text

    emacs emacs emacs EMACS emacs

and move from word to word with "*", point will move to "EMACS" if
`case-fold-search' is t. If `search-upper-case' is non-nil, the
upper-case characters may enable a case-sensitive search. We make
this dependent on the previous command: if it was "*", then the old
search is continued; if it was something else, a new search is
started, which may be case-sensitive.