Change `integerp' condition in `vimpulse-keypress-parser'
In XEmacs, (integerp ?\C-0) => nil and (characterp ?\C-0) => t;
in GNU Emacs, (integerp ?\C-0) => t and (characterp ?\C-0) => nil
(provided `characterp' exists, otherwise we alias it to `integerp').
Since the `-'/`logand' arithmetic for converting keys to integers
works the same in both cases, we change the condition to
(or (characterp char) (integerp char)).