Add DEBUG parameter to tests and suites
This provides verbose output useful on a per-test basis. Normally,
tests are run in "batch mode", where the test code is wrapped in a
protective `condition-case' form. This is usually what we want, since
it doesn't halt code execution when vimpulse-test.el is loaded in
.emacs. Thus, if something fails, we will get a warning, but Emacs
will still load to the best of its ability.

Otherwise, in "debug mode", the debugger is entered on error and
execution is halted. This mode is enabled by running a test or suite
interactively, or by passing t to the DEBUG parameter, e.g., (test t).
The "debug mode" sets `debug-on-error' to t.

Incidentally, the DEBUG parameter is also used for selecting among the
different "clauses" of a test function mentioned in the message of
commit 67edb0a. In the "internal" cases, DEBUG is either `batch' or
`debug', and the test silently returns t if it passes or returns an
error message string if it fails (or it enters the debugger). When
tests are run as part of a suite, they are run in this way. The suite
collects all error messages and presents them to the user.

In the "user" cases, DEBUG is either nil (the default) or t, and the
test looks for a suite to present it. The test calls a suite which in
turn runs the test "internally". In sketch:

    +--------+    +-------------------+    +---------------+
    | (test) | => | (suite ... 'test) | => | (test 'batch) |
    +--------+    +-------------------+    +---------------+

If DEBUG is nil (the default when called programmatically), the test
is ultimately called in "batch mode". If DEBUG is t (the default when
called interactively), the test is ultimately called in "debug mode".
A :debug keyword argument may be specified to let a test be called in
"debug mode" by default, but this should be used carefully.