Make block insert and append repeatable
It is now possible to repeat actions like "C-v j I <text>" with ".".
This does away with Alessandro's `vimpulse-visual-insert-coords' code,
which was based on the number of lines; the new code uses markers
instead. We also remove the redefinition of `viper-exit-insert-state',
since the "update all lines when we hit ESC" behavior can be
conveniently fitted into `viper-vi-state-hook'.