Keep isearch overlays when `vimpulse-flash-delay' expires after "n"/"N"
After a "n"/"N" command, clean isearch overlays only upon the
execution of a subsequent "non-search" command, not the expiration of
`vimpulse-flash-delay'. This is so that the invisible area opened by
"n"/"N" will not become permanently open as the result of the
expiration of `vimpulse-flash-delay'. In other words, the opened area
can still be closed by a subsequent "n"/"N" when the match leaves the
area, and the only way to make the area "permanently open" is to
execute a "non-search" command after the search (any key-press other
than "n" or "N").